Move ScalarMaybeUndef into `value.rs`
